+++ packages/diblob-visualizer/src/lib/index.ts
@@ -4,7 +4,7 @@
  * Main exports for embedding the visualizer in other applications
  */
 
-export * from './container-introspection';
+export * from '../shared/container-introspection';
 export { default as DependencyGraph } from './DependencyGraph.svelte';
 export { default as DiblobVisualizer } from './DiblobVisualizer.svelte';
 export { default as RemoteDiblobVisualizer } from './RemoteDiblobVisualizer.svelte';